frost_core
Function challenge

pub fn challenge<C>(
    R: &Element<C>,
    verifying_key: &Element<C>,
    msg: &[u8],
) -> Challenge<C>
where
    C: Ciphersuite,

Available on crate feature internals only.

Generates the challenge as is required for Schnorr signatures.

Deals in bytes, so that FROST and singleton signing and verification can use it with different types.

This is the only invocation of the H2 hash function from the RFC.