frost_core::frost
Module keys

FROST keys, keygen, key shares

Modules

dkg

Distributed Key Generation functions and structures.

repairable

Repairable Threshold Scheme

Structs

CoefficientCommitment

A Group::Element newtype that is a commitment to one coefficient of our secret polynomial.

KeyPackage

A FROST keypair, which can be generated either by a trusted dealer or using a DKG.

PublicKeyPackage

Public data that contains all the signers’ verifying shares as well as the group verifying key.

SecretShare

A secret share generated by performing a (t-out-of-n) secret sharing scheme, generated by a dealer performing generate_with_dealer.

SigningShare

A secret scalar value representing a signer’s share of the group secret.

VerifiableSecretSharingCommitment

Contains the commitments to the coefficients for our secret polynomial f, used to generate participants’ key shares.

VerifyingShare

A public group element that represents a single signer’s public verification share.

Enums

IdentifierList

The identifier list to use when generating key shares.

Functions

generate_with_dealer

Allows all participants’ keys to be generated using a central, trusted dealer.

reconstruct

Recompute the secret from at least min_signers secret shares (inside KeyPackages) using Lagrange interpolation.

split

Splits an existing key into FROST shares.