frost_core::frost::keys
Module dkg

Distributed Key Generation functions and structures.

The DKG module supports generating FROST key shares in a distributed manner, without a trusted dealer, via two rounds of communication between all participants.

This implements FROST KeyGen from the original FROST paper, specifically Figure 1. This protocol is a variant of Pedersen’s DKG that additionally requires each participant to demonstrate knowledge of their secret by providing other participants with proof in zero knowledge, instantiated as a Schnorr signature, to protect against rogue-key attacks in the setting where t ≥ n/2.

In Pedersen’s DKG, each of the n participants executes Feldman’s Verifiable Secret Sharing (VSS) as the dealer in parallel, and derives their secret share as the sum of the shares received from each of the n VSS executions.

As required for any multi-party protocol using Feldman’s VSS, the key generation stage in FROST requires participants to maintain a consistent view of the pubic commitments to the secret polynomial coefficients. This DKG protocol requires participants to broadcast the commitment values honestly (e.g., participants do not provide different commitment values to a subset of participants) over a secure broadcast channel.

For more details and an example, see the ciphersuite-specific crates, e.g. frost_ristretto255::keys::dkg.

Modules

round1

DKG Round 1 structures.

round2

DKG Round 2 structures.

Functions

part1

Performs the first part of the distributed key generation protocol for the given participant.

part2

Performs the second part of the distributed key generation protocol for the participant holding the given round1::SecretPackage, given the received round1::Packages received from the other participants.

part3

Performs the third and final part of the distributed key generation protocol for the participant holding the given round2::SecretPackage, given the received round1::Packages and round2::Packages received from the other participants.