frost_core::frost::keys::dkg
Function part1

pub fn part1<C: Ciphersuite, R: RngCore + CryptoRng>(
    identifier: Identifier<C>,
    max_signers: u16,
    min_signers: u16,
    rng: R,
) -> Result<(SecretPackage<C>, Package<C>), Error<C>>

Performs the first part of the distributed key generation protocol for the given participant.

It returns the round1::SecretPackage that must be kept in memory by the participant for the other steps, and the round1::Package that must be sent to other participants.