frost_core::frost::keys::dkg
Function part2

pub fn part2<C: Ciphersuite>(
    secret_package: SecretPackage<C>,
    round1_packages: &HashMap<Identifier<C>, Package<C>>,
) -> Result<(SecretPackage<C>, HashMap<Identifier<C>, Package<C>>), Error<C>>

Performs the second part of the distributed key generation protocol for the participant holding the given round1::SecretPackage, given the received round1::Packages received from the other participants.

round1_packages maps the identifier of each participant to the round1::Package they sent. These identifiers must come from whatever mapping the coordinator has between communication channels and participants, i.e. they must have assurance that the round1::Package came from the participant with that identifier.

It returns the round2::SecretPackage that must be kept in memory by the participant for the final step, and the a map of round2::Packages that must be sent to each participant who has the given identifier in the map key.