frost_core::frost::keys::dkg
Function part3

pub fn part3<C: Ciphersuite>(
    round2_secret_package: &SecretPackage<C>,
    round1_packages: &HashMap<Identifier<C>, Package<C>>,
    round2_packages: &HashMap<Identifier<C>, Package<C>>,
) -> Result<(KeyPackage<C>, PublicKeyPackage<C>), Error<C>>

Performs the third and final part of the distributed key generation protocol for the participant holding the given round2::SecretPackage, given the received round1::Packages and round2::Packages received from the other participants.

round1_packages must be the same used in part2().

round2_packages maps the identifier of each participant to the round2::Package they sent. These identifiers must come from whatever mapping the coordinator has between communication channels and participants, i.e. they must have assurance that the round2::Package came from the participant with that identifier.

It returns the KeyPackage that has the long-lived key share for the participant, and the PublicKeyPackages that has public information about all participants; both of which are required to compute FROST signatures.