frost_core::frost::keys
Function generate_with_dealer

pub fn generate_with_dealer<C: Ciphersuite, R: RngCore + CryptoRng>(
    max_signers: u16,
    min_signers: u16,
    identifiers: IdentifierList<'_, C>,
    rng: &mut R,
) -> Result<(HashMap<Identifier<C>, SecretShare<C>>, PublicKeyPackage<C>), Error<C>>

Allows all participants’ keys to be generated using a central, trusted dealer.

Under the hood, this performs verifiable secret sharing, which itself uses Shamir secret sharing, from which each share becomes a participant’s secret key. The output from this function is a set of shares along with one single commitment that participants use to verify the integrity of the share.

Implements trusted_dealer_keygen from the spec.