frost_core::frost::keys
Function split

pub fn split<C: Ciphersuite, R: RngCore + CryptoRng>(
    key: &SigningKey<C>,
    max_signers: u16,
    min_signers: u16,
    identifiers: IdentifierList<'_, C>,
    rng: &mut R,
) -> Result<(HashMap<Identifier<C>, SecretShare<C>>, PublicKeyPackage<C>), Error<C>>

Splits an existing key into FROST shares.

This is identical to generate_with_dealer but receives an existing key instead of generating a fresh one. This is useful in scenarios where the key needs to be generated externally or must be derived from e.g. a seed phrase.