frost_core::frost::keys::repairable
Function repair_share_step_1

pub fn repair_share_step_1<C: Ciphersuite, R: RngCore + CryptoRng>(
    helpers: &[Identifier<C>],
    share_i: &SecretShare<C>,
    rng: &mut R,
    participant: Identifier<C>,
) -> Result<HashMap<Identifier<C>, Scalar<C>>, Error<C>>

Step 1 of RTS.

Generates the “delta” values from helper_i to help participant recover their share where helpers contains the identifiers of all the helpers (including helper_i), and share_i is the share of helper_i.

Returns a HashMap mapping which value should be sent to which participant.