frost_core::frost
Module round1

FROST Round 1 functionality and types

Structs

GroupCommitmentShare

One signer’s share of the group commitment, derived from their individual signing commitments and the binding factor rho.

Nonce

A scalar that is a signing nonce.

NonceCommitment

A group element that is a commitment to a signing nonce share.

SigningCommitments

Published by each participant in the first round of the signing protocol.

SigningNonces

Comprised of hiding and binding nonces.

Functions

commit

Performed once by each participant selected for the signing operation.

preprocess

Done once by each participant, to generate their nonces and commitments that are then used during signing.