frost_core::frost::round1
Function commit

pub fn commit<C, R>(
    secret: &SigningShare<C>,
    rng: &mut R,
) -> (SigningNonces<C>, SigningCommitments<C>)
where
    C: Ciphersuite,
    R: CryptoRng + RngCore,

Performed once by each participant selected for the signing operation.

Implements commit from the spec.

Generates the signing nonces and commitments to be used in the signing operation.