frost_core::frost::round1
Function preprocess

pub fn preprocess<C, R>(
    num_nonces: u8,
    secret: &SigningShare<C>,
    rng: &mut R,
) -> (Vec<SigningNonces<C>>, Vec<SigningCommitments<C>>)
where
    C: Ciphersuite,
    R: CryptoRng + RngCore,

Done once by each participant, to generate their nonces and commitments that are then used during signing.

This is only needed if pre-processing is needed (for 1-round FROST). For regular 2-round FROST, use commit.

When performing signing using two rounds, num_nonces would equal 1, to perform the first round. Batching entails generating more than one nonce/commitment pair at a time. Nonces should be stored in secret storage for later use, whereas the commitments are published.