frost_core::frost::round2
Function sign

pub fn sign<C: Ciphersuite>(
    signing_package: &SigningPackage<C>,
    signer_nonces: &SigningNonces<C>,
    key_package: &KeyPackage<C>,
) -> Result<SignatureShare<C>, Error<C>>

Performed once by each participant selected for the signing operation.

Implements sign from the spec.

Receives the message to be signed and a set of signing commitments and a set of randomizing commitments to be used in that signing operation, including that for this participant.

Assumes the participant has already determined which nonce corresponds with the commitment that was assigned by the coordinator in the SigningPackage.