frost_core
Module frost

An implementation of FROST (Flexible Round-Optimized Schnorr Threshold) signatures.

For key generation, refer to the keys module. For round-specific types and functions, refer to the round1 and round2 modules.

This module contains types and functions not directly related to key generation and the FROST rounds.

Modules

keys

FROST keys, keygen, key shares

round1

FROST Round 1 functionality and types

round2

FROST Round 2 functionality and types, for signature share generation

Structs

BindingFactor

The binding factor, also known as rho (ρ)

BindingFactorList

A list of binding factors and their associated identifiers.

GroupCommitment

The product of all signers’ individual commitments, published as part of the final signature.

Identifier

A FROST participant identifier.

SigningPackage

Generated by the coordinator of the signing operation and distributed to each signing party

Functions

aggregate

Aggregates the signature shares to produce a final signature that can be verified with the group public key.

compute_binding_factor_listinternals

compute_binding_factors in the spec

compute_group_commitmentinternals

Generates the group commitment which is published as part of the joint Schnorr signature.

compute_lagrange_coefficientinternals

Generates a lagrange coefficient.

derive_interpolating_valueinternals

Generates the lagrange coefficient for the i’th participant (for signer_id).